lambda-ingest/requirements.txt
-i https://pypi.org/simple
aiohttp==3.9.5; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.4.0; python_version >= '3.8'
async-timeout==4.0.3; python_version < '3.11'
attrs==23.2.0; python_version >= '3.7'
beautifulsoup4==4.12.3; python_full_version >= '3.6.0'
boto3==1.34.146; python_version >= '3.8'
botocore==1.34.146; python_version >= '3.8'
certifi==2024.7.4; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
dataclasses-json==0.6.7; python_version >= '3.7' and python_version < '4.0'
deprecated==1.2.14; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
dirtyjson==1.0.8
distro==1.9.0; python_version >= '3.6'
dnspython==2.6.1; python_version >= '3.8'
exceptiongroup==1.2.2; python_version < '3.11'
frozenlist==1.4.1; python_version >= '3.8'
fsspec==2024.6.1; python_version >= '3.8'
greenlet==3.0.3; python_version < '3.13' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.5; python_version >= '3.8'
httpx==0.27.0; python_version >= '3.8'
idna==3.7; python_version >= '3.5'
jmespath==1.0.1; python_version >= '3.7'
joblib==1.4.2; python_version >= '3.8'
llama-cloud==0.0.9; python_version >= '3.8' and python_version < '4'
llama-index==0.10.57; python_version < '4.0' and python_full_version >= '3.8.1'
llama-index-agent-openai==0.2.9; python_version < '4.0' and python_full_version >= '3.8.1'
llama-index-cli==0.1.12; python_version < '4.0' and python_full_version >= '3.8.1'
llama-index-core==0.10.57; python_version < '4.0' and python_full_version >= '3.8.1'
llama-index-embeddings-openai==0.1.11; python_version < '4.0' and python_full_version >= '3.8.1'
llama-index-indices-managed-llama-cloud==0.2.5; python_version < '4.0' and python_full_version >= '3.8.1'
llama-index-legacy==0.9.48; python_version < '4.0' and python_full_version >= '3.8.1'
llama-index-llms-openai==0.1.27; python_version < '4.0' and python_full_version >= '3.8.1'
llama-index-multi-modal-llms-openai==0.1.8; python_version < '4.0' and python_full_version >= '3.8.1'
llama-index-program-openai==0.1.7; python_version < '4.0' and python_full_version >= '3.8.1'
llama-index-question-gen-openai==0.1.3; python_version < '4.0' and python_full_version >= '3.8.1'
llama-index-readers-file==0.1.30; python_version < '4.0' and python_full_version >= '3.8.1'
llama-index-readers-llama-parse==0.1.6; python_version < '4.0' and python_full_version >= '3.8.1'
llama-index-vector-stores-awsdocdb==0.1.6; python_version < '4.0' and python_full_version >= '3.8.1'
llama-parse==0.4.9; python_version < '4.0' and python_full_version >= '3.8.1'
marshmallow==3.21.3; python_version >= '3.8'
multidict==6.0.5; python_version >= '3.7'
mypy-extensions==1.0.0; python_version >= '3.5'
nest-asyncio==1.6.0; python_version >= '3.5'
networkx==3.3; python_version >= '3.10'
nltk==3.8.1; python_version >= '3.7'
numpy==1.26.4; python_version >= '3.9'
openai==1.37.0; python_full_version >= '3.7.1'
packaging==24.1; python_version >= '3.8'
pandas==2.2.2; python_version >= '3.9'
pillow==10.4.0; python_version >= '3.8'
pydantic==2.8.2; python_version >= '3.8'
pydantic-core==2.20.1; python_version >= '3.8'
pymongo==4.8.0; python_version >= '3.8'
pypdf==4.3.1; python_version >= '3.6'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.1
pyyaml==6.0.1; python_version >= '3.6'
regex==2024.5.15; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
s3transfer==0.10.2; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
soupsieve==2.5; python_version >= '3.8'
sqlalchemy[asyncio]==2.0.31; python_version >= '3.7'
striprtf==0.0.26
tenacity==8.5.0; python_version >= '3.8'
tiktoken==0.7.0; python_version >= '3.8'
tqdm==4.66.4; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
typing-inspect==0.9.0
tzdata==2024.1; python_version >= '2'
urllib3==2.2.2; python_version >= '3.10'
wrapt==1.16.0; python_version >= '3.6'
yarl==1.9.4; python_version >= '3.7'
